generic array support and type casting

+++ test/integration_tests/test_pytorch.py
@@ -29,7 +29,7 @@ class PytorchTest(unittest.TestCase):
         perform_query(query)
 
         create_udf_query = """CREATE UDF FastRCNNObjectDetector
-                  INPUT  (Frame_Array NDARRAY (3, 256, 256))
+                  INPUT  (Frame_Array NDARRAY UINT8(3, 256, 256))
                   OUTPUT (label TEXT(10))
                   TYPE  Classification
                   IMPL  'src/udfs/fastrcnn_object_detector.py';
@@ -47,7 +47,7 @@ class PytorchTest(unittest.TestCase):
         perform_query(query)
 
         create_udf_query = """CREATE UDF SSDObjectDetector
-                  INPUT  (Frame_Array NDARRAY (3, 256, 256))
+                  INPUT  (Frame_Array NDARRAY UINT8(3, 256, 256))
                   OUTPUT (label TEXT(10))
                   TYPE  Classification
                   IMPL  'src/udfs/ssd_object_detector.py';
